refactor: move validateKeyboardLayout to a seperate file
