+++ src/app.ts
@@ -1,28 +1,12 @@
 import express, { Request, Response } from 'express'
 import gode, { EngLayout, ThaLayout } from 'gode.js'
 import cors from 'cors'
+import * as utils from './utils'
 
 const app = express()
 const port = 3000
 
 app.use(cors())
-
-function validateKeyboardLayout(englayout: string, thalayout: string) {
-    const validThaLayout = ['Manoonchai', 'Kedmanee']
-    const validEngLayout = ['QWERTY']
-    let isThaValid = false
-    let isEngValid = false
-    if (validThaLayout.includes(thalayout)) {
-        isThaValid = true
-    }
-    if (validEngLayout.includes(englayout)) {
-        isEngValid = true
-    }
-    return {
-        isThaValid: isThaValid,
-        isEngValid: isEngValid
-    }
-}
 
 app.get('/', (req: Request, res: Response) => {
     res.redirect('https://github.com/godeProject/godeAPI')
@@ -57,7 +41,7 @@ app.get('/v1/getans/', (req: Request, res: Response) => {
 app.get('/v2/convert/englayout/:englayout/thalayout/:thalayout/', (req: Request, res: Response) => {
     try {
         let parameter = req.params
-        let validity = validateKeyboardLayout(parameter.englayout as string, parameter.thalayout as string)
+        let validity = utils.validateKeyboardLayout(parameter.englayout as string, parameter.thalayout as string)
         let tl: unknown = parameter.thalayout
         let el: unknown = parameter.englayout
         let message: unknown = req.query.message
